requirements.txt
spotipy == 2.22.1
python-dotenv == 1.0.0
beautifulsoup4 == 4.12.2